CassetteNote/backend_simple/requirements.txt
flask
flask-cors
flask-jwt-extended
werkzeug
pillow
